refactors viz into modules
