+++ epilepsy12/common_view_functions/render_charts/__init__.py
@@ -1,0 +1,3 @@
+from .pie_chart import render_pie_pct_passed_for_kpi_agg
+from .bar_chart import render_bar_pct_passed_for_kpi_agg
+from .chart_html import ChartHTML
